refactor : user 패키지 생성
seller와 customer 패키지를 통합하여 user 패키지 생성
