+++ src/main/java/com/itonse/clotheslink/admin/service/TokenService.java
@@ -1,8 +1,8 @@
 package com.itonse.clotheslink.admin.service;
 
 import com.itonse.clotheslink.common.UserVo;
-import com.itonse.clotheslink.customer.domain.Customer;
-import com.itonse.clotheslink.seller.domain.Seller;
+import com.itonse.clotheslink.user.domain.Customer;
+import com.itonse.clotheslink.user.domain.Seller;
 
 public interface TokenService {
