--- src/main/java/com/itonse/clotheslink/customer/dto/SignInDto.java
@@ -1,16 +1,0 @@
-package com.itonse.clotheslink.customer.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-@Getter
-public class SignInDto {
-
-    private String email;
-    private String password;
-}
