--- src/main/java/com/itonse/clotheslink/customer/service/CustomerAuthService.java
@@ -1,13 +1,0 @@
-package com.itonse.clotheslink.customer.service;
-
-import com.itonse.clotheslink.customer.dto.SignInDto;
-import com.itonse.clotheslink.customer.dto.SignUpDto;
-import com.itonse.clotheslink.customer.dto.UserInfoResponse;
-
-public interface CustomerAuthService {
-
-    UserInfoResponse signUp(SignUpDto dto);
-
-    String signIn(SignInDto dto);
-
-}
