+++ src/main/java/com/itonse/clotheslink/product/service/ProductService.java
@@ -4,7 +4,7 @@ import com.itonse.clotheslink.product.domain.Product;
 import com.itonse.clotheslink.product.dto.ProductDto;
 import com.itonse.clotheslink.product.dto.ProductSummaryInfo;
 import com.itonse.clotheslink.product.dto.UpdateProductDto;
-import com.itonse.clotheslink.seller.domain.Seller;
+import com.itonse.clotheslink.user.domain.Seller;
 
 public interface ProductService {
     ProductSummaryInfo addProduct(String token, ProductDto dto);
