--- src/main/java/com/itonse/clotheslink/seller/dto/UserInfoResponse.java
@@ -1,15 +1,0 @@
-package com.itonse.clotheslink.seller.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-@Builder
-public class UserInfoResponse {
-    private Long id;
-    private String email;
-}
