--- src/main/java/com/itonse/clotheslink/seller/service/SellerAuthService.java
@@ -1,12 +1,0 @@
-package com.itonse.clotheslink.seller.service;
-
-import com.itonse.clotheslink.seller.dto.SignInDto;
-import com.itonse.clotheslink.seller.dto.UserInfoResponse;
-import com.itonse.clotheslink.seller.dto.SignUpDto;
-
-public interface SellerAuthService {
-    UserInfoResponse signUp(SignUpDto dto);
-
-    String signin(SignInDto dto);
-
-}
