+++ src/main/java/com/itonse/clotheslink/user/dto/SignUpDto.java
@@ -1,6 +1,7 @@
-package com.itonse.clotheslink.customer.dto;
+package com.itonse.clotheslink.user.dto;
 
-import com.itonse.clotheslink.customer.domain.Customer;
+import com.itonse.clotheslink.user.domain.Customer;
+import com.itonse.clotheslink.user.domain.Seller;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -12,14 +13,20 @@ import lombok.NoArgsConstructor;
 @Getter
 public class SignUpDto {
     private String email;
-    private String name;
     private String password;
     private String phone;
+
+    public static Seller toSellerEntity(SignUpDto dto) {
+        return Seller.builder()
+                .email(dto.getEmail())
+                .password(dto.getPassword())
+                .phone(dto.getPhone())
+                .build();
+    }
 
     public static Customer toCustomerEntity(SignUpDto dto) {
         return Customer.builder()
                 .email(dto.getEmail())
-                .name(dto.getName())
                 .password(dto.getPassword())
                 .phone(dto.getPhone())
                 .build();
