+++ src/main/java/com/itonse/clotheslink/user/dto/SignUpForm.java
@@ -1,4 +1,4 @@
-package com.itonse.clotheslink.seller.dto;
+package com.itonse.clotheslink.user.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -13,20 +13,24 @@ import javax.validation.constraints.Size;
 @NoArgsConstructor
 @Builder
 @Getter
-public class SignInForm {
+public class SignUpForm {
 
     @Email(message = "이메일 형식이 아닙니다.")
     @NotBlank(message = "이메일은 필수 항목입니다.")
     private String email;
 
-    @Size(min = 8, max = 20, message = "비밀번호는 8자 이상으로 20자 이하로 입력해주세요.")
+    @Size(min = 8, max = 20, message = "비밀번호는 최소 8자 ~ 최대 20자 이내로 입력해주세요.")
     @NotBlank(message = "비밀번호는 필수 항목입니다.")
     private String password;
 
-    public static SignInDto toSignInDto(SignInForm form) {
-        return SignInDto.builder()
+    @NotBlank(message = "연락처는 필수 항목 입니다.")
+    private String phone;
+
+    public static SignUpDto toSignUpDto(SignUpForm form) {
+        return SignUpDto.builder()
                 .email(form.getEmail())
                 .password(form.getPassword())
+                .phone(form.getPhone())
                 .build();
     }
 }
