+++ src/main/java/com/itonse/clotheslink/user/dto/UserInfoResponse.java
@@ -1,4 +1,4 @@
-package com.itonse.clotheslink.seller.dto;
+package com.itonse.clotheslink.user.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -7,10 +7,9 @@ import lombok.NoArgsConstructor;
 
 @AllArgsConstructor
 @NoArgsConstructor
-@Builder
 @Getter
-public class SignInDto {
-
+@Builder
+public class UserInfoResponse {
+    private Long id;
     private String email;
-    private String password;
 }
