+++ src/main/java/com/itonse/clotheslink/user/repository/SellerRepository.java
@@ -1,6 +1,6 @@
-package com.itonse.clotheslink.seller.repository;
+package com.itonse.clotheslink.user.repository;
 
-import com.itonse.clotheslink.seller.domain.Seller;
+import com.itonse.clotheslink.user.domain.Seller;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -8,6 +8,7 @@ import java.util.Optional;
 
 @Repository
 public interface SellerRepository extends JpaRepository<Seller, Long> {
+
     Optional<Seller> findByEmail(String email);
 
     Optional<Seller> findByEmailAndPassword(String email, String password);
