+++ src/main/java/com/itonse/clotheslink/user/service/CustomerAuthService.java
@@ -1,0 +1,12 @@
+package com.itonse.clotheslink.user.service;
+
+import com.itonse.clotheslink.user.dto.SignInDto;
+import com.itonse.clotheslink.user.dto.SignUpDto;
+import com.itonse.clotheslink.user.dto.UserInfoResponse;
+
+public interface CustomerAuthService {
+
+    UserInfoResponse signUp(SignUpDto dto);
+
+    String signIn(SignInDto dto);
+}
